pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pymatgraph"
version = "0.3.0"
description = "A Python package for high-performance GPU/CPU buffer rendering with support for tables, text, and graphics."
readme = "README.md"
license = "MIT"
authors = [
    { name="Alec Candidato", email="aleccandidato@gmail.com" }
]
requires-python = ">=3.8"
dependencies = ["torch", "numpy", "Pillow", "pygame", "kivy"]
dynamic = ["classifiers"]
include = ["matrixbuffer/fonts/*"]

[project.urls]
Homepage = "https://github.com/0202alcc/pymatgraph"
Documentation = "https://github.com/0202alcc/pymatgraph#readme"
Repository = "https://github.com/0202alcc/pymatgraph"
